Change story and passage events behaviour
Now `start` event occurs before first render of start passage and `show` event occurs after every passage render (also the first one).

Now the whole passage object is passed to event callback (not only the name) and also an `isStart` argument (true or false if the passage is the first one or not).